# Add a skip constructor to the stream type