This is a work in progress PR for fixing https://github.com/composewell/streamly/issues/90

The major operations in `vector` for which the Skip constructor exists like `filter`, `drop` etc they need to be modified still and the whole thing needs benchmarking. 

Also does the CPS encoding need to change as well to allow transformation between the two alternate representations?