@harendra-kumar 2018-10-27
I would suggest that you try it on just a couple of relevant operations and comment out the rest of the operations. That way you do not have to make all the changes before knowing whether it has the desired impact.