@Abhiroop 2018-10-30
These numbers are on an Intel Core i5 Processor. The comparisons are shown against `vector` which is able to leverage fusion and is the fastest among other streaming libraries.

## Benchmarks with Skip

```
Composed Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark   vector(μs)(base) streamly(%)(-base)
----------- ---------------- ------------------
filter-scan        137522.15            +193.90
filter-drop          7027.87            +178.62
drop-scan          327172.69             +30.32
take-scan          374211.60             +11.06
filter-take          1367.04              +0.64
take-map             1383.34              +0.01
filter-map            819.87              -2.05
scan-map           478217.01              -5.32
drop-map             7124.62             -70.77
take-drop           32724.59             -89.48
```

```
Filtering  Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark       vector(μs)(base) streamly(%)(-base)
--------------- ---------------- ------------------
dropWhile-true           1959.22              +4.10
takeWhile-true            656.18              +3.92
filter-even               651.32              +0.37
filter-all-in             685.80              +0.34
dropWhile-false         31877.75              +0.25
filter-all-out            660.73              -1.29
take-all                 1351.15              -1.45
drop-all                  981.48             -31.71
drop-one                 6897.82             -70.01
```
## Benchmarks without Skip (streamly-0.5.2)

```
Composed Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark   vector(μs)(base) streamly(%)(-base)
----------- ---------------- ------------------
filter-take          1394.25            +995.77
filter-drop          6296.87            +113.55
drop-map             6978.72             +65.00
filter-map            663.91             +50.27
take-map             1346.74              +2.16
take-drop           30819.03              +1.24
filter-scan        139686.84             -48.89
drop-scan          391854.90             -65.63
take-scan          369179.51             -81.59
scan-map           466626.51             -85.28
```

```
Filtering  Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark       vector(μs)(base) streamly(%)(-base)
--------------- ---------------- ------------------
filter-even               677.98            +202.54
drop-all                 1044.65            +129.21
dropWhile-true           2037.21             +69.83
drop-one                 6917.37             +48.94
filter-all-in             691.37             +44.53
filter-all-out            693.56              +0.95
take-all                 1347.89              -0.18
takeWhile-true            709.55              -2.77
dropWhile-false         33145.88             -66.36
```