@Abhiroop 2018-10-31
`filter-drop` with Skip was previously 178% slower than `vector`. A new implementation of `drop` as implemented in `vector` improves it considerably as observed by @pranaysashank .

It improves `drop-all` as well. Few functions like `take-drop` shows minor regressions(compared to `streamly-0.5.2`) but are still faster than `vector`

```
Composed Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark   vector(μs)(base) streamly(%)(-base)
----------- ---------------- ------------------
filter-scan        160595.40            +143.09
drop-scan          364036.78             +26.61
take-scan          366582.61             +13.10
take-map             1379.57              +0.25
take-drop           30197.05              -0.02
drop-map             6728.82              -0.18
filter-map            698.35              -1.78
filter-drop          6335.64              -3.97
filter-take          1501.01              -5.21
scan-map           495187.06              -9.46
```

```
Filtering  Operations x 4 (Lower is Better)(time)(Median)(Diff using min estimator)
Benchmark       vector(μs)(base) streamly(%)(-base)
--------------- ---------------- ------------------
filter-all-out            713.36              +3.13
drop-all                 1002.23              +2.77
drop-one                 7014.66              +2.09
filter-even               685.09              +2.00
take-all                 1387.22              +0.75
dropWhile-true           2149.84              -0.55
filter-all-in             712.41              -2.05
dropWhile-false         37086.43              -4.63
takeWhile-true            709.23              -5.55
```